Remove new/assignment-approved from the ReqMgr2/WMStats list of active workflows
